Adding tests for SyntaxtNodeExtensions.
