+++ tests/D2L.CodeStyle.Analysis.UnitTests/MutabilityInspectorTests.cs
@@ -7,53 +7,9 @@ using NUnit.Framework;
 namespace D2L.CodeStyle.Analysis {
 
 	[TestFixture]
-	public class MutabilityInspectorTests {
+	public class MutabilityInspectorTests : TestBase {
 
 		private readonly MutabilityInspector m_inspector = new MutabilityInspector();
-
-		private static CSharpCompilation Compile( string source ) {
-			var tree = CSharpSyntaxTree.ParseText( source );
-			var compilation = CSharpCompilation.Create(
-				assemblyName: "TestAssembly",
-				syntaxTrees: new[] { tree },
-				references: new[] {
-					MetadataReference.CreateFromFile( typeof( object ).Assembly.Location ),
-					MetadataReference.CreateFromFile( typeof( ImmutableArray ).Assembly.Location )
-				}
-			);
-			return compilation;
-		}
-
-		private ITypeSymbol Type( string text ) {
-			var source = $"namespace D2L {{ {text} }}";
-			var compilation = Compile( source );
-
-			var toReturn = compilation.GetSymbolsWithName(
-				predicate: n => true,
-				filter: SymbolFilter.Type
-			).OfType<ITypeSymbol>().FirstOrDefault();
-			Assert.IsNotNull( toReturn );
-			Assert.AreNotEqual( TypeKind.Error, toReturn.TypeKind );
-			return toReturn;
-		}
-
-		private IFieldSymbol Field( string text ) {
-			var type = Type( "sealed class Fake { " + text + "; }" );
-
-			var toReturn = type.GetMembers().OfType<IFieldSymbol>().FirstOrDefault();
-			Assert.IsNotNull( toReturn );
-			Assert.AreNotEqual( TypeKind.Error, toReturn.Type.TypeKind );
-			return toReturn;
-		}
-
-		private IPropertySymbol Property( string text ) {
-			var type = Type( "sealed class Fake { " + text + "; }" );
-
-			var toReturn = type.GetMembers().OfType<IPropertySymbol>().FirstOrDefault();
-			Assert.IsNotNull( toReturn );
-			Assert.AreNotEqual( TypeKind.Error, toReturn.Type.TypeKind );
-			return toReturn;
-		}
 
 		[Test]
 		public void IsFieldMutable_Private_True() {
